[Fix]: Fixed a bug with the XP pump upgrade messing up the stored levels

+++ src/main/java/net/p3pp3rf1y/sophisticatedcore/upgrades/xppump/XpPumpUpgradeWrapper.java
@@ -113,7 +113,7 @@ public class XpPumpUpgradeWrapper extends UpgradeWrapperBase<XpPumpUpgradeWrappe
             outer.commit();
 
             if (!drained.isEmpty()) {
-                player.giveExperiencePoints((int) XpHelper.liquidToExperience((int) drained.getAmount()));
+                player.giveExperiencePoints((int) XpHelper.liquidToExperience(drained.getAmount()));
             }
         }
 	}
